test(bigquery): refactor integration tests
This PR is a first cut at organizing the integration tests a bit
more.  The existing single integration_test.go was getting somewhat
long and unwieldy.

This PR neither changes any tests nor adds new test coverage.
