+++ bigquery/integration_test.go
@@ -31,7 +31,6 @@ import (
 	"cloud.google.com/go/civil"
 	datacatalog "cloud.google.com/go/datacatalog/apiv1"
 	"cloud.google.com/go/httpreplay"
-	"cloud.google.com/go/iam"
 	"cloud.google.com/go/internal"
 	"cloud.google.com/go/internal/pretty"
 	"cloud.google.com/go/internal/testutil"
@@ -323,216 +322,6 @@ func TestIntegration_JobFrom(t *testing.T) {
 
 }
 
-func TestIntegration_TableCreate(t *testing.T) {
-	// Check that creating a record field with an empty schema is an error.
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	table := dataset.Table("t_bad")
-	schema := Schema{
-		{Name: "rec", Type: RecordFieldType, Schema: Schema{}},
-	}
-	err := table.Create(context.Background(), &TableMetadata{
-		Schema:         schema,
-		ExpirationTime: testTableExpiration.Add(5 * time.Minute),
-	})
-	if err == nil {
-		t.Fatal("want error, got nil")
-	}
-	if !hasStatusCode(err, http.StatusBadRequest) {
-		t.Fatalf("want a 400 error, got %v", err)
-	}
-}
-
-func TestIntegration_TableCreateWithConstraints(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	table := dataset.Table("constraints")
-	schema := Schema{
-		{Name: "str_col", Type: StringFieldType, MaxLength: 10},
-		{Name: "bytes_col", Type: BytesFieldType, MaxLength: 150},
-		{Name: "num_col", Type: NumericFieldType, Precision: 20},
-		{Name: "bignumeric_col", Type: BigNumericFieldType, Precision: 30, Scale: 5},
-	}
-	err := table.Create(context.Background(), &TableMetadata{
-		Schema:         schema,
-		ExpirationTime: testTableExpiration.Add(5 * time.Minute),
-	})
-	if err != nil {
-		t.Fatalf("table create error: %v", err)
-	}
-
-	meta, err := table.Metadata(context.Background())
-	if err != nil {
-		t.Fatalf("couldn't get metadata: %v", err)
-	}
-
-	if diff := testutil.Diff(meta.Schema, schema); diff != "" {
-		t.Fatalf("got=-, want=+:\n%s", diff)
-	}
-
-}
-
-func TestIntegration_TableCreateView(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	tableIdentifier, _ := table.Identifier(StandardSQLID)
-	defer table.Delete(ctx)
-
-	// Test that standard SQL views work.
-	view := dataset.Table("t_view_standardsql")
-	query := fmt.Sprintf("SELECT APPROX_COUNT_DISTINCT(name) FROM %s", tableIdentifier)
-	err := view.Create(context.Background(), &TableMetadata{
-		ViewQuery:      query,
-		UseStandardSQL: true,
-	})
-	if err != nil {
-		t.Fatalf("table.create: Did not expect an error, got: %v", err)
-	}
-	if err := view.Delete(ctx); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestIntegration_TableMetadata(t *testing.T) {
-
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-	// Check table metadata.
-	md, err := table.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// TODO(jba): check md more thorougly.
-	if got, want := md.FullID, fmt.Sprintf("%s:%s.%s", dataset.ProjectID, dataset.DatasetID, table.TableID); got != want {
-		t.Errorf("metadata.FullID: got %q, want %q", got, want)
-	}
-	if got, want := md.Type, RegularTable; got != want {
-		t.Errorf("metadata.Type: got %v, want %v", got, want)
-	}
-	if got, want := md.ExpirationTime, testTableExpiration; !got.Equal(want) {
-		t.Errorf("metadata.Type: got %v, want %v", got, want)
-	}
-
-	// Check that timePartitioning is nil by default
-	if md.TimePartitioning != nil {
-		t.Errorf("metadata.TimePartitioning: got %v, want %v", md.TimePartitioning, nil)
-	}
-
-	// Create tables that have time partitioning
-	partitionCases := []struct {
-		timePartitioning TimePartitioning
-		wantExpiration   time.Duration
-		wantField        string
-		wantPruneFilter  bool
-	}{
-		{TimePartitioning{}, time.Duration(0), "", false},
-		{TimePartitioning{Expiration: time.Second}, time.Second, "", false},
-		{TimePartitioning{RequirePartitionFilter: true}, time.Duration(0), "", true},
-		{
-			TimePartitioning{
-				Expiration:             time.Second,
-				Field:                  "date",
-				RequirePartitionFilter: true,
-			}, time.Second, "date", true},
-	}
-
-	schema2 := Schema{
-		{Name: "name", Type: StringFieldType},
-		{Name: "date", Type: DateFieldType},
-	}
-
-	clustering := &Clustering{
-		Fields: []string{"name"},
-	}
-
-	// Currently, clustering depends on partitioning.  Interleave testing of the two features.
-	for i, c := range partitionCases {
-		table := dataset.Table(fmt.Sprintf("t_metadata_partition_nocluster_%v", i))
-		clusterTable := dataset.Table(fmt.Sprintf("t_metadata_partition_cluster_%v", i))
-
-		// Create unclustered, partitioned variant and get metadata.
-		err = table.Create(context.Background(), &TableMetadata{
-			Schema:           schema2,
-			TimePartitioning: &c.timePartitioning,
-			ExpirationTime:   testTableExpiration,
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-		defer table.Delete(ctx)
-		md, err := table.Metadata(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		// Created clustered table and get metadata.
-		err = clusterTable.Create(context.Background(), &TableMetadata{
-			Schema:           schema2,
-			TimePartitioning: &c.timePartitioning,
-			ExpirationTime:   testTableExpiration,
-			Clustering:       clustering,
-		})
-		if err != nil {
-			t.Fatal(err)
-		}
-		clusterMD, err := clusterTable.Metadata(ctx)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		for _, v := range []*TableMetadata{md, clusterMD} {
-			got := v.TimePartitioning
-			want := &TimePartitioning{
-				Type:                   DayPartitioningType,
-				Expiration:             c.wantExpiration,
-				Field:                  c.wantField,
-				RequirePartitionFilter: c.wantPruneFilter,
-			}
-			if !testutil.Equal(got, want) {
-				t.Errorf("metadata.TimePartitioning: got %v, want %v", got, want)
-			}
-			// Manipulate RequirePartitionFilter at the table level.
-			mdUpdate := TableMetadataToUpdate{
-				RequirePartitionFilter: !want.RequirePartitionFilter,
-			}
-
-			newmd, err := table.Update(ctx, mdUpdate, "")
-			if err != nil {
-				t.Errorf("failed to invert RequirePartitionFilter on %s: %v", table.FullyQualifiedName(), err)
-			}
-			if newmd.RequirePartitionFilter == want.RequirePartitionFilter {
-				t.Errorf("inverting table-level RequirePartitionFilter on %s failed, want %t got %t", table.FullyQualifiedName(), !want.RequirePartitionFilter, newmd.RequirePartitionFilter)
-			}
-			// Also verify that the clone of RequirePartitionFilter in the TimePartitioning message is consistent.
-			if newmd.RequirePartitionFilter != newmd.TimePartitioning.RequirePartitionFilter {
-				t.Errorf("inconsistent RequirePartitionFilter.  Table: %t, TimePartitioning: %t", newmd.RequirePartitionFilter, newmd.TimePartitioning.RequirePartitionFilter)
-			}
-
-		}
-
-		if md.Clustering != nil {
-			t.Errorf("metadata.Clustering was not nil on unclustered table %s", table.TableID)
-		}
-		got := clusterMD.Clustering
-		want := clustering
-		if clusterMD.Clustering != clustering {
-			if !testutil.Equal(got, want) {
-				t.Errorf("metadata.Clustering: got %v, want %v", got, want)
-			}
-		}
-	}
-
-}
-
 func TestIntegration_SnapshotAndRestore(t *testing.T) {
 
 	if client == nil {
@@ -757,371 +546,6 @@ func TestIntegration_RemoveTimePartitioning(t *testing.T) {
 	}
 }
 
-func TestIntegration_DatasetCreate(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	ds := client.Dataset(datasetIDs.New())
-	wmd := &DatasetMetadata{Name: "name", Location: "EU"}
-	err := ds.Create(ctx, wmd)
-	if err != nil {
-		t.Fatal(err)
-	}
-	gmd, err := ds.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := gmd.Name, wmd.Name; got != want {
-		t.Errorf("name: got %q, want %q", got, want)
-	}
-	if got, want := gmd.Location, wmd.Location; got != want {
-		t.Errorf("location: got %q, want %q", got, want)
-	}
-	if err := ds.Delete(ctx); err != nil {
-		t.Fatalf("deleting dataset %v: %v", ds, err)
-	}
-}
-
-func TestIntegration_DatasetMetadata(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	md, err := dataset.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := md.FullID, fmt.Sprintf("%s:%s", dataset.ProjectID, dataset.DatasetID); got != want {
-		t.Errorf("FullID: got %q, want %q", got, want)
-	}
-	jan2016 := time.Date(2016, 1, 1, 0, 0, 0, 0, time.UTC)
-	if md.CreationTime.Before(jan2016) {
-		t.Errorf("CreationTime: got %s, want > 2016-1-1", md.CreationTime)
-	}
-	if md.LastModifiedTime.Before(jan2016) {
-		t.Errorf("LastModifiedTime: got %s, want > 2016-1-1", md.LastModifiedTime)
-	}
-
-	// Verify that we get a NotFound for a nonexistent dataset.
-	_, err = client.Dataset("does_not_exist").Metadata(ctx)
-	if err == nil || !hasStatusCode(err, http.StatusNotFound) {
-		t.Errorf("got %v, want NotFound error", err)
-	}
-}
-
-func TestIntegration_DatasetDelete(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	ds := client.Dataset(datasetIDs.New())
-	if err := ds.Create(ctx, nil); err != nil {
-		t.Fatalf("creating dataset %s: %v", ds.DatasetID, err)
-	}
-	if err := ds.Delete(ctx); err != nil {
-		t.Fatalf("deleting dataset %s: %v", ds.DatasetID, err)
-	}
-}
-
-func TestIntegration_DatasetDeleteWithContents(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	ds := client.Dataset(datasetIDs.New())
-	if err := ds.Create(ctx, nil); err != nil {
-		t.Fatalf("creating dataset %s: %v", ds.DatasetID, err)
-	}
-	table := ds.Table(tableIDs.New())
-	if err := table.Create(ctx, nil); err != nil {
-		t.Fatalf("creating table %s in dataset %s: %v", table.TableID, table.DatasetID, err)
-	}
-	// We expect failure here
-	if err := ds.Delete(ctx); err == nil {
-		t.Fatalf("non-recursive delete of dataset %s succeeded unexpectedly.", ds.DatasetID)
-	}
-	if err := ds.DeleteWithContents(ctx); err != nil {
-		t.Fatalf("deleting recursively dataset %s: %v", ds.DatasetID, err)
-	}
-}
-
-func TestIntegration_DatasetUpdateETags(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-
-	check := func(md *DatasetMetadata, wantDesc, wantName string) {
-		if md.Description != wantDesc {
-			t.Errorf("description: got %q, want %q", md.Description, wantDesc)
-		}
-		if md.Name != wantName {
-			t.Errorf("name: got %q, want %q", md.Name, wantName)
-		}
-	}
-
-	ctx := context.Background()
-	md, err := dataset.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if md.ETag == "" {
-		t.Fatal("empty ETag")
-	}
-	// Write without ETag succeeds.
-	desc := md.Description + "d2"
-	name := md.Name + "n2"
-	md2, err := dataset.Update(ctx, DatasetMetadataToUpdate{Description: desc, Name: name}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	check(md2, desc, name)
-
-	// Write with original ETag fails because of intervening write.
-	_, err = dataset.Update(ctx, DatasetMetadataToUpdate{Description: "d", Name: "n"}, md.ETag)
-	if err == nil {
-		t.Fatal("got nil, want error")
-	}
-
-	// Write with most recent ETag succeeds.
-	md3, err := dataset.Update(ctx, DatasetMetadataToUpdate{Description: "", Name: ""}, md2.ETag)
-	if err != nil {
-		t.Fatal(err)
-	}
-	check(md3, "", "")
-}
-
-func TestIntegration_DatasetUpdateDefaultExpiration(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	_, err := dataset.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Set the default expiration time.
-	md, err := dataset.Update(ctx, DatasetMetadataToUpdate{DefaultTableExpiration: time.Hour}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if md.DefaultTableExpiration != time.Hour {
-		t.Fatalf("got %s, want 1h", md.DefaultTableExpiration)
-	}
-	// Omitting DefaultTableExpiration doesn't change it.
-	md, err = dataset.Update(ctx, DatasetMetadataToUpdate{Name: "xyz"}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if md.DefaultTableExpiration != time.Hour {
-		t.Fatalf("got %s, want 1h", md.DefaultTableExpiration)
-	}
-	// Setting it to 0 deletes it (which looks like a 0 duration).
-	md, err = dataset.Update(ctx, DatasetMetadataToUpdate{DefaultTableExpiration: time.Duration(0)}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if md.DefaultTableExpiration != 0 {
-		t.Fatalf("got %s, want 0", md.DefaultTableExpiration)
-	}
-}
-
-func TestIntegration_DatasetUpdateAccess(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	md, err := dataset.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Create a sample UDF so we can verify adding authorized UDFs
-	routineID := routineIDs.New()
-	routine := dataset.Routine(routineID)
-	routineSQLID, _ := routine.Identifier(StandardSQLID)
-
-	sql := fmt.Sprintf(`
-			CREATE FUNCTION %s(x INT64) AS (x * 3);`,
-		routineSQLID)
-	if _, _, err := runQuerySQL(ctx, sql); err != nil {
-		t.Fatal(err)
-	}
-	defer routine.Delete(ctx)
-
-	origAccess := append([]*AccessEntry(nil), md.Access...)
-	newEntries := []*AccessEntry{
-		{
-			Role:       ReaderRole,
-			Entity:     "Joe@example.com",
-			EntityType: UserEmailEntity,
-		},
-		{
-			Role:       ReaderRole,
-			Entity:     "allUsers",
-			EntityType: IAMMemberEntity,
-		},
-		{
-			EntityType: RoutineEntity,
-			Routine:    routine,
-		},
-		{
-			EntityType: DatasetEntity,
-			Dataset: &DatasetAccessEntry{
-				Dataset:     otherDataset,
-				TargetTypes: []string{"VIEWS"},
-			},
-		},
-	}
-
-	newAccess := append(md.Access, newEntries...)
-	dm := DatasetMetadataToUpdate{Access: newAccess}
-	md, err = dataset.Update(ctx, dm, md.ETag)
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer func() {
-		_, err := dataset.Update(ctx, DatasetMetadataToUpdate{Access: origAccess}, md.ETag)
-		if err != nil {
-			t.Log("could not restore dataset access list")
-		}
-	}()
-
-	if diff := testutil.Diff(md.Access, newAccess, cmpopts.SortSlices(lessAccessEntries), cmpopts.IgnoreUnexported(Routine{}, Dataset{})); diff != "" {
-		t.Errorf("got=-, want=+:\n%s", diff)
-	}
-}
-
-// Comparison function for AccessEntries to enable order insensitive equality checking.
-func lessAccessEntries(x, y *AccessEntry) bool {
-	if x.Entity < y.Entity {
-		return true
-	}
-	if x.Entity > y.Entity {
-		return false
-	}
-	if x.EntityType < y.EntityType {
-		return true
-	}
-	if x.EntityType > y.EntityType {
-		return false
-	}
-	if x.Role < y.Role {
-		return true
-	}
-	if x.Role > y.Role {
-		return false
-	}
-	if x.View == nil {
-		return y.View != nil
-	}
-	if x.Routine == nil {
-		return y.Routine == nil
-	}
-	if x.Dataset == nil {
-		return y.Dataset == nil
-	}
-	return false
-}
-
-func TestIntegration_DatasetUpdateLabels(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	_, err := dataset.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var dm DatasetMetadataToUpdate
-	dm.SetLabel("label", "value")
-	md, err := dataset.Update(ctx, dm, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := md.Labels["label"], "value"; got != want {
-		t.Errorf("got %q, want %q", got, want)
-	}
-	dm = DatasetMetadataToUpdate{}
-	dm.DeleteLabel("label")
-	md, err = dataset.Update(ctx, dm, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, ok := md.Labels["label"]; ok {
-		t.Error("label still present after deletion")
-	}
-}
-
-func TestIntegration_TableUpdateLabels(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-
-	var tm TableMetadataToUpdate
-	tm.SetLabel("label", "value")
-	md, err := table.Update(ctx, tm, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got, want := md.Labels["label"], "value"; got != want {
-		t.Errorf("got %q, want %q", got, want)
-	}
-	tm = TableMetadataToUpdate{}
-	tm.DeleteLabel("label")
-	md, err = table.Update(ctx, tm, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, ok := md.Labels["label"]; ok {
-		t.Error("label still present after deletion")
-	}
-}
-
-func TestIntegration_Tables(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-	wantName := table.FullyQualifiedName()
-
-	// This test is flaky due to eventual consistency.
-	ctx, cancel := context.WithTimeout(ctx, 10*time.Second)
-	defer cancel()
-	err := internal.Retry(ctx, gax.Backoff{}, func() (stop bool, err error) {
-		// Iterate over tables in the dataset.
-		it := dataset.Tables(ctx)
-		var tableNames []string
-		for {
-			tbl, err := it.Next()
-			if err == iterator.Done {
-				break
-			}
-			if err != nil {
-				return false, err
-			}
-			tableNames = append(tableNames, tbl.FullyQualifiedName())
-		}
-		// Other tests may be running with this dataset, so there might be more
-		// than just our table in the list. So don't try for an exact match; just
-		// make sure that our table is there somewhere.
-		for _, tn := range tableNames {
-			if tn == wantName {
-				return true, nil
-			}
-		}
-		return false, fmt.Errorf("got %v\nwant %s in the list", tableNames, wantName)
-	})
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
 // setupPolicyTag is a helper for setting up policy tags in the datacatalog service.
 //
 // It returns a string for a policy tag identifier and a cleanup function, or an error.
@@ -1202,66 +626,6 @@ func TestIntegration_ColumnACLs(t *testing.T) {
 		Description: "foo",
 	}); err != nil {
 		t.Errorf("failed to create new table with policy tag: %v", err)
-	}
-}
-
-func TestIntegration_TableIAM(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-
-	// Check to confirm some of our default permissions.
-	checkedPerms := []string{"bigquery.tables.get",
-		"bigquery.tables.getData", "bigquery.tables.update"}
-	perms, err := table.IAM().TestPermissions(ctx, checkedPerms)
-	if err != nil {
-		t.Fatalf("IAM().TestPermissions: %v", err)
-	}
-	if len(perms) != len(checkedPerms) {
-		t.Errorf("mismatch in permissions, got (%s) wanted (%s)", strings.Join(perms, " "), strings.Join(checkedPerms, " "))
-	}
-
-	// Get existing policy, add a binding for all authenticated users.
-	policy, err := table.IAM().Policy(ctx)
-	if err != nil {
-		t.Fatalf("IAM().Policy: %v", err)
-	}
-	wantedRole := iam.RoleName("roles/bigquery.dataViewer")
-	wantedMember := "allAuthenticatedUsers"
-	policy.Add(wantedMember, wantedRole)
-	if err := table.IAM().SetPolicy(ctx, policy); err != nil {
-		t.Fatalf("IAM().SetPolicy: %v", err)
-	}
-
-	// Verify policy mutations were persisted by refetching policy.
-	updatedPolicy, err := table.IAM().Policy(ctx)
-	if err != nil {
-		t.Fatalf("IAM.Policy (after update): %v", err)
-	}
-	foundRole := false
-	for _, r := range updatedPolicy.Roles() {
-		if r == wantedRole {
-			foundRole = true
-			break
-		}
-	}
-	if !foundRole {
-		t.Errorf("Did not find added role %s in the set of %d roles.",
-			wantedRole, len(updatedPolicy.Roles()))
-	}
-	members := updatedPolicy.Members(wantedRole)
-	foundMember := false
-	for _, m := range members {
-		if m == wantedMember {
-			foundMember = true
-			break
-		}
-	}
-	if !foundMember {
-		t.Errorf("Did not find member %s in role %s", wantedMember, wantedRole)
 	}
 }
 
@@ -1860,122 +1224,6 @@ func testInsertAndReadNullable(t *testing.T, ts testStructNullable, wantRow []Va
 	}
 	if diff := testutil.Diff(sn, want, roundToMicros); diff != "" {
 		t.Error(diff)
-	}
-}
-
-func TestIntegration_TableUpdate(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-
-	// Test Update of non-schema fields.
-	tm, err := table.Metadata(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-	wantDescription := tm.Description + "more"
-	wantName := tm.Name + "more"
-	wantExpiration := tm.ExpirationTime.Add(time.Hour * 24)
-	got, err := table.Update(ctx, TableMetadataToUpdate{
-		Description:    wantDescription,
-		Name:           wantName,
-		ExpirationTime: wantExpiration,
-	}, tm.ETag)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if got.Description != wantDescription {
-		t.Errorf("Description: got %q, want %q", got.Description, wantDescription)
-	}
-	if got.Name != wantName {
-		t.Errorf("Name: got %q, want %q", got.Name, wantName)
-	}
-	if got.ExpirationTime != wantExpiration {
-		t.Errorf("ExpirationTime: got %q, want %q", got.ExpirationTime, wantExpiration)
-	}
-	if !testutil.Equal(got.Schema, schema) {
-		t.Errorf("Schema: got %v, want %v", pretty.Value(got.Schema), pretty.Value(schema))
-	}
-
-	// Blind write succeeds.
-	_, err = table.Update(ctx, TableMetadataToUpdate{Name: "x"}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	// Write with old etag fails.
-	_, err = table.Update(ctx, TableMetadataToUpdate{Name: "y"}, got.ETag)
-	if err == nil {
-		t.Fatal("Update with old ETag succeeded, wanted failure")
-	}
-
-	// Test schema update.
-	// Columns can be added. schema2 is the same as schema, except for the
-	// added column in the middle.
-	nested := Schema{
-		{Name: "nested", Type: BooleanFieldType},
-		{Name: "other", Type: StringFieldType},
-	}
-	schema2 := Schema{
-		schema[0],
-		{Name: "rec2", Type: RecordFieldType, Schema: nested},
-		schema[1],
-		schema[2],
-	}
-
-	got, err = table.Update(ctx, TableMetadataToUpdate{Schema: schema2}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Wherever you add the column, it appears at the end.
-	schema3 := Schema{schema2[0], schema2[2], schema2[3], schema2[1]}
-	if !testutil.Equal(got.Schema, schema3) {
-		t.Errorf("add field:\ngot  %v\nwant %v",
-			pretty.Value(got.Schema), pretty.Value(schema3))
-	}
-
-	// Updating with the empty schema succeeds, but is a no-op.
-	got, err = table.Update(ctx, TableMetadataToUpdate{Schema: Schema{}}, "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !testutil.Equal(got.Schema, schema3) {
-		t.Errorf("empty schema:\ngot  %v\nwant %v",
-			pretty.Value(got.Schema), pretty.Value(schema3))
-	}
-
-	// Error cases when updating schema.
-	for _, test := range []struct {
-		desc   string
-		fields Schema
-	}{
-		{"change from optional to required", Schema{
-			{Name: "name", Type: StringFieldType, Required: true},
-			schema3[1],
-			schema3[2],
-			schema3[3],
-		}},
-		{"add a required field", Schema{
-			schema3[0], schema3[1], schema3[2], schema3[3],
-			{Name: "req", Type: StringFieldType, Required: true},
-		}},
-		{"remove a field", Schema{schema3[0], schema3[1], schema3[2]}},
-		{"remove a nested field", Schema{
-			schema3[0], schema3[1], schema3[2],
-			{Name: "rec2", Type: RecordFieldType, Schema: Schema{nested[0]}}}},
-		{"remove all nested fields", Schema{
-			schema3[0], schema3[1], schema3[2],
-			{Name: "rec2", Type: RecordFieldType, Schema: Schema{}}}},
-	} {
-		_, err = table.Update(ctx, TableMetadataToUpdate{Schema: Schema(test.fields)}, "")
-		if err == nil {
-			t.Errorf("%s: want error, got nil", test.desc)
-		} else if !hasStatusCode(err, 400) {
-			t.Errorf("%s: want 400, got %v", test.desc, err)
-		}
 	}
 }
 
@@ -2877,32 +2125,6 @@ func TestIntegration_QueryUseLegacySQL(t *testing.T) {
 	}
 }
 
-func TestIntegration_TableUseLegacySQL(t *testing.T) {
-	// Test UseLegacySQL and UseStandardSQL for Table.Create.
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-	table := newTable(t, schema)
-	defer table.Delete(ctx)
-	for i, test := range useLegacySQLTests {
-		view := dataset.Table(fmt.Sprintf("t_view_%d", i))
-		tm := &TableMetadata{
-			ViewQuery:      fmt.Sprintf("SELECT word from %s", test.t),
-			UseStandardSQL: test.std,
-			UseLegacySQL:   test.legacy,
-		}
-		err := view.Create(ctx, tm)
-		gotErr := err != nil
-		if gotErr && !test.err {
-			t.Errorf("%+v:\nunexpected error: %v", test, err)
-		} else if !gotErr && test.err {
-			t.Errorf("%+v:\nsucceeded, but want error", test)
-		}
-		_ = view.Delete(ctx)
-	}
-}
-
 func TestIntegration_ListJobs(t *testing.T) {
 	// It's difficult to test the list of jobs, because we can't easily
 	// control what's in it. Also, there are many jobs in the test project,
@@ -3365,215 +2587,6 @@ func TestIntegration_ModelLifecycle(t *testing.T) {
 	// Delete the model.
 	if err := model.Delete(ctx); err != nil {
 		t.Fatalf("failed to delete model: %v", err)
-	}
-}
-
-func TestIntegration_RoutineScalarUDF(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-
-	// Create a scalar UDF routine via API.
-	routineID := routineIDs.New()
-	routine := dataset.Routine(routineID)
-	err := routine.Create(ctx, &RoutineMetadata{
-		Type:     "SCALAR_FUNCTION",
-		Language: "SQL",
-		Body:     "x * 3",
-		Arguments: []*RoutineArgument{
-			{
-				Name: "x",
-				DataType: &StandardSQLDataType{
-					TypeKind: "INT64",
-				},
-			},
-		},
-	})
-	if err != nil {
-		t.Fatalf("Create: %v", err)
-	}
-}
-
-func TestIntegration_RoutineJSUDF(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-
-	// Create a scalar UDF routine via API.
-	routineID := routineIDs.New()
-	routine := dataset.Routine(routineID)
-	meta := &RoutineMetadata{
-		Language: "JAVASCRIPT", Type: "SCALAR_FUNCTION",
-		Description:      "capitalizes using javascript",
-		DeterminismLevel: Deterministic,
-		Arguments: []*RoutineArgument{
-			{Name: "instr", Kind: "FIXED_TYPE", DataType: &StandardSQLDataType{TypeKind: "STRING"}},
-		},
-		ReturnType: &StandardSQLDataType{TypeKind: "STRING"},
-		Body:       "return instr.toUpperCase();",
-	}
-	if err := routine.Create(ctx, meta); err != nil {
-		t.Fatalf("Create: %v", err)
-	}
-
-	newMeta := &RoutineMetadataToUpdate{
-		Language:    meta.Language,
-		Body:        meta.Body,
-		Arguments:   meta.Arguments,
-		Description: meta.Description,
-		ReturnType:  meta.ReturnType,
-		Type:        meta.Type,
-
-		DeterminismLevel: NotDeterministic,
-	}
-	if _, err := routine.Update(ctx, newMeta, ""); err != nil {
-		t.Fatalf("Update: %v", err)
-	}
-}
-
-func TestIntegration_RoutineComplexTypes(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-
-	routineID := routineIDs.New()
-	routine := dataset.Routine(routineID)
-	routineSQLID, _ := routine.Identifier(StandardSQLID)
-	sql := fmt.Sprintf(`
-		CREATE FUNCTION %s(
-			arr ARRAY<STRUCT<name STRING, val INT64>>
-		  ) AS (
-			  (SELECT SUM(IF(elem.name = "foo",elem.val,null)) FROM UNNEST(arr) AS elem)
-		  )`,
-		routineSQLID)
-	if _, _, err := runQuerySQL(ctx, sql); err != nil {
-		t.Fatal(err)
-	}
-	defer routine.Delete(ctx)
-
-	meta, err := routine.Metadata(ctx)
-	if err != nil {
-		t.Fatalf("Metadata: %v", err)
-	}
-	if meta.Type != "SCALAR_FUNCTION" {
-		t.Fatalf("routine type mismatch, got %s want SCALAR_FUNCTION", meta.Type)
-	}
-	if meta.Language != "SQL" {
-		t.Fatalf("language type mismatch, got  %s want SQL", meta.Language)
-	}
-	want := []*RoutineArgument{
-		{
-			Name: "arr",
-			DataType: &StandardSQLDataType{
-				TypeKind: "ARRAY",
-				ArrayElementType: &StandardSQLDataType{
-					TypeKind: "STRUCT",
-					StructType: &StandardSQLStructType{
-						Fields: []*StandardSQLField{
-							{
-								Name: "name",
-								Type: &StandardSQLDataType{
-									TypeKind: "STRING",
-								},
-							},
-							{
-								Name: "val",
-								Type: &StandardSQLDataType{
-									TypeKind: "INT64",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-	if diff := testutil.Diff(meta.Arguments, want); diff != "" {
-		t.Fatalf("%+v: -got, +want:\n%s", meta.Arguments, diff)
-	}
-}
-
-func TestIntegration_RoutineLifecycle(t *testing.T) {
-	if client == nil {
-		t.Skip("Integration tests skipped")
-	}
-	ctx := context.Background()
-
-	// Create a scalar UDF routine via a CREATE FUNCTION query
-	routineID := routineIDs.New()
-	routine := dataset.Routine(routineID)
-	routineSQLID, _ := routine.Identifier(StandardSQLID)
-
-	sql := fmt.Sprintf(`
-		CREATE FUNCTION %s(x INT64) AS (x * 3);`,
-		routineSQLID)
-	if _, _, err := runQuerySQL(ctx, sql); err != nil {
-		t.Fatal(err)
-	}
-	defer routine.Delete(ctx)
-
-	// Get the routine metadata.
-	curMeta, err := routine.Metadata(ctx)
-	if err != nil {
-		t.Fatalf("couldn't get metadata: %v", err)
-	}
-
-	want := "SCALAR_FUNCTION"
-	if curMeta.Type != want {
-		t.Errorf("Routine type mismatch.  got %s want %s", curMeta.Type, want)
-	}
-
-	want = "SQL"
-	if curMeta.Language != want {
-		t.Errorf("Language mismatch. got %s want %s", curMeta.Language, want)
-	}
-
-	// Perform an update to change the routine body and description.
-	want = "x * 4"
-	wantDescription := "an updated description"
-	// during beta, update doesn't allow partial updates.  Provide all fields.
-	newMeta, err := routine.Update(ctx, &RoutineMetadataToUpdate{
-		Body:        want,
-		Arguments:   curMeta.Arguments,
-		Description: wantDescription,
-		ReturnType:  curMeta.ReturnType,
-		Type:        curMeta.Type,
-	}, curMeta.ETag)
-	if err != nil {
-		t.Fatalf("Update: %v", err)
-	}
-	if newMeta.Body != want {
-		t.Fatalf("Update body failed. want %s got %s", want, newMeta.Body)
-	}
-	if newMeta.Description != wantDescription {
-		t.Fatalf("Update description failed. want %s got %s", wantDescription, newMeta.Description)
-	}
-
-	// Ensure presence when enumerating the model list.
-	it := dataset.Routines(ctx)
-	seen := false
-	for {
-		r, err := it.Next()
-		if err == iterator.Done {
-			break
-		}
-		if err != nil {
-			t.Fatal(err)
-		}
-		if r.RoutineID == routineID {
-			seen = true
-		}
-	}
-	if !seen {
-		t.Fatal("routine not listed in dataset")
-	}
-
-	// Delete the model.
-	if err := routine.Delete(ctx); err != nil {
-		t.Fatalf("failed to delete routine: %v", err)
 	}
 }
